Update README.md
Mercado Flask is a simple web application built with the Flask framework, which simulates a basic management system for a virtual market. The application allows users to register, with encrypted passwords for security, and view products registered in the system. The project uses SQLite as a database to store information about users and products. Features: User registration: Users can register with a username, email and password, with validations to avoid duplication. Product display: Products are listed on the product page, with information such as name, price, barcode and description. Security: User passwords are encrypted using the Flask-Bcrypt library, ensuring greater system security. Database: Uses Flask-SQLAlchemy to manage data persistence in the SQLite database.